+++ test/e2e/conftest.py
@@ -2,28 +2,13 @@ from pathlib import Path
 
 import pytest
 from fastapi.testclient import TestClient
-from run import generate_app
-from syrupy.assertion import SnapshotAssertion
-from syrupy.extensions.json import JSONSnapshotExtension
 
+from run import generate_app
 from voicevox_engine.core_initializer import initialize_cores
 from voicevox_engine.preset import PresetManager
 from voicevox_engine.setting import SettingLoader
 from voicevox_engine.tts_pipeline import make_tts_engines_from_cores
 from voicevox_engine.utility.core_version_utility import get_latest_core_version
-
-
-@pytest.fixture
-def snapshot_json(snapshot: SnapshotAssertion) -> SnapshotAssertion:
-    """
-    syrupyでJSONをsnapshotするためのfixture。
-
-    Examples
-    --------
-    >>> def test_foo(snapshot_json: JSONSnapshotExtension):
-    >>>     assert snapshot_json == {"key": "value"}
-    """
-    return snapshot.use_extension(JSONSnapshotExtension)
 
 
 @pytest.fixture(scope="session")
