+++ test/test_tts_engine.py
@@ -1,8 +1,10 @@
+import json
 from typing import Union
 from unittest import TestCase
 from unittest.mock import Mock
 
 import numpy
+from pydantic.json import pydantic_encoder
 from syrupy.extensions.json import JSONSnapshotExtension
 
 from voicevox_engine.dev.core.mock import MockCoreWrapper
@@ -27,8 +29,6 @@ from voicevox_engine.tts_pipeline.tts_engine import (
     to_flatten_phonemes,
     unvoiced_mora_phoneme_list,
 )
-
-from .e2e.conftest import snapshot_json  # noqa: F401
 
 TRUE_NUM_PHONEME = 45
 
@@ -707,13 +707,11 @@ class TestTTSEngine(TestCase):
         self.assertEqual(result, true_result)
 
 
-def test_mocked_update_length_output(
-    snapshot_json: JSONSnapshotExtension,  # noqa: F811
-) -> None:
+def test_mocked_update_length_output(snapshot_json: JSONSnapshotExtension) -> None:
     # Inputs
     tts_engine = TTSEngine(MockCoreWrapper())
     hello_hiho = _gen_hello_hiho_accent_phrases()
     # Outputs
     result = tts_engine.update_length(hello_hiho, StyleId(1))
     # Tests
-    assert snapshot_json == result  # type: ignore
+    assert snapshot_json == json.loads(json.dumps(result, default=pydantic_encoder))
